move capture to ui
